Store times as Long types
closes #119

+++ app/src/main/java/com/enaboapps/switchify/screens/settings/models/SettlingsScreenModel.kt
@@ -7,28 +7,28 @@ import com.enaboapps.switchify.preferences.PreferenceManager
 class SettingsScreenModel(context: Context) : ViewModel() {
     private val preferenceManager = PreferenceManager(context)
 
-    fun getScanRate(): Int {
-        return preferenceManager.getIntegerValue(PreferenceManager.Keys.PREFERENCE_KEY_SCAN_RATE)
+    fun getScanRate(): Long {
+        return preferenceManager.getLongValue(PreferenceManager.Keys.PREFERENCE_KEY_SCAN_RATE)
     }
 
-    fun setScanRate(rate: Int) {
-        preferenceManager.setIntegerValue(PreferenceManager.Keys.PREFERENCE_KEY_SCAN_RATE, rate)
+    fun setScanRate(rate: Long) {
+        preferenceManager.setLongValue(PreferenceManager.Keys.PREFERENCE_KEY_SCAN_RATE, rate)
     }
 
-    fun getRefineScanRate(): Int {
-        return preferenceManager.getIntegerValue(PreferenceManager.Keys.PREFERENCE_KEY_REFINE_SCAN_RATE)
+    fun getRefineScanRate(): Long {
+        return preferenceManager.getLongValue(PreferenceManager.Keys.PREFERENCE_KEY_REFINE_SCAN_RATE)
     }
 
-    fun setRefineScanRate(rate: Int) {
-        preferenceManager.setIntegerValue(PreferenceManager.Keys.PREFERENCE_KEY_REFINE_SCAN_RATE, rate)
+    fun setRefineScanRate(rate: Long) {
+        preferenceManager.setLongValue(PreferenceManager.Keys.PREFERENCE_KEY_REFINE_SCAN_RATE, rate)
     }
 
-    fun getSwitchHoldTime(): Int {
-        return preferenceManager.getIntegerValue(PreferenceManager.Keys.PREFERENCE_KEY_SWITCH_HOLD_TIME)
+    fun getSwitchHoldTime(): Long {
+        return preferenceManager.getLongValue(PreferenceManager.Keys.PREFERENCE_KEY_SWITCH_HOLD_TIME)
     }
 
-    fun setSwitchHoldTime(time: Int) {
-        preferenceManager.setIntegerValue(PreferenceManager.Keys.PREFERENCE_KEY_SWITCH_HOLD_TIME, time)
+    fun setSwitchHoldTime(time: Long) {
+        preferenceManager.setLongValue(PreferenceManager.Keys.PREFERENCE_KEY_SWITCH_HOLD_TIME, time)
     }
 
     fun getAutoSelect(): Boolean {
@@ -36,14 +36,20 @@ class SettingsScreenModel(context: Context) : ViewModel() {
     }
 
     fun setAutoSelect(autoSelect: Boolean) {
-        preferenceManager.setBooleanValue(PreferenceManager.Keys.PREFERENCE_KEY_AUTO_SELECT, autoSelect)
+        preferenceManager.setBooleanValue(
+            PreferenceManager.Keys.PREFERENCE_KEY_AUTO_SELECT,
+            autoSelect
+        )
     }
 
-    fun getAutoSelectDelay(): Int {
-        return preferenceManager.getIntegerValue(PreferenceManager.Keys.PREFERENCE_KEY_AUTO_SELECT_DELAY)
+    fun getAutoSelectDelay(): Long {
+        return preferenceManager.getLongValue(PreferenceManager.Keys.PREFERENCE_KEY_AUTO_SELECT_DELAY)
     }
 
-    fun setAutoSelectDelay(delay: Int) {
-        preferenceManager.setIntegerValue(PreferenceManager.Keys.PREFERENCE_KEY_AUTO_SELECT_DELAY, delay)
+    fun setAutoSelectDelay(delay: Long) {
+        preferenceManager.setLongValue(
+            PreferenceManager.Keys.PREFERENCE_KEY_AUTO_SELECT_DELAY,
+            delay
+        )
     }
 }
